fix(server): isolate native workspace indexing from the event loop
