+++ apps/server/src/workspace/WorkspaceSearchIndex.test.ts
@@ -1,348 +1,227 @@
-import {
-  FileFinder,
-  type FileItem,
-  type GrepCursor,
-  type GrepOptions,
-  type GrepResult,
-} from "@ff-labs/fff-node";
-import { afterEach, expect, it } from "@effect/vitest";
-import * as Cause from "effect/Cause";
+// @effect-diagnostics nodeBuiltinImport:off
+import * as NodeNet from "node:net";
+import * as NodeHttp from "node:http";
+import * as NodeHttpClient from "@effect/platform-node/NodeHttpClient";
+import * as HttpClient from "effect/unstable/http/HttpClient";
+import { expect, it } from "@effect/vitest";
 import * as Effect from "effect/Effect";
+import * as Deferred from "effect/Deferred";
+import * as Fiber from "effect/Fiber";
+import * as Scope from "effect/Scope";
 import * as Exit from "effect/Exit";
-import { vi } from "vite-plus/test";
+import * as TestClock from "effect/testing/TestClock";
+import { HostProcessEnvironment } from "@t3tools/shared/hostProcess";
+import * as Index from "./WorkspaceSearchIndex.ts";
+import { WorkspaceSearchWorkerPath } from "./WorkspaceSearchProcess.ts";
+import { WorkspaceSearchHost } from "./WorkspaceSearchHost.ts";
 
-import * as WorkspaceSearchIndex from "./WorkspaceSearchIndex.ts";
+const mockWorker = new URL("../../scripts/workspace-search-mock.ts", import.meta.url);
+const withWorker = <A, E, R>(effect: Effect.Effect<A, E, R>) =>
+  effect.pipe(
+    Effect.provide(WorkspaceSearchHost.layer),
+    Effect.provideService(WorkspaceSearchWorkerPath, mockWorker),
+  );
 
-afterEach(() => {
-  vi.restoreAllMocks();
+const receipts = Effect.fn(function* () {
+  const blocked = yield* Deferred.make<void>();
+  const exited = yield* Deferred.make<void>();
+  const server = NodeNet.createServer((socket) => {
+    socket.once("data", () => Deferred.doneUnsafe(blocked, Effect.void));
+    socket.once("close", () => Deferred.doneUnsafe(exited, Effect.void));
+  });
+  yield* Effect.acquireRelease(
+    Effect.promise(() => new Promise<void>((resolve) => server.listen(0, "127.0.0.1", resolve))),
+    () => Effect.promise(() => new Promise<void>((resolve) => server.close(() => resolve()))),
+  );
+  const address = server.address();
+  if (!address || typeof address === "string") return yield* Effect.die("Missing receipt port");
+  return {
+    blocked,
+    exited,
+    environment: { ...process.env, T3_SEARCH_TEST_RECEIPT_PORT: String(address.port) },
+  };
 });
 
-function fileItem(relativePath: string): FileItem {
-  return {
-    relativePath,
-    fileName: relativePath.slice(relativePath.lastIndexOf("/") + 1),
-    size: 1,
-    modified: 0,
-    accessFrecencyScore: 0,
-    modificationFrecencyScore: 0,
-    totalFrecencyScore: 0,
-    gitStatus: "clean",
-  };
-}
+it.effect(
+  "keeps HTTP responsive during a blocked search and rebuilds shared indexes after timeout",
+  () =>
+    Effect.scoped(
+      Effect.gen(function* () {
+        const control = yield* receipts();
+        const server = NodeHttp.createServer((_request, response) => response.end("healthy"));
+        yield* Effect.acquireRelease(
+          Effect.promise(
+            () => new Promise<void>((resolve) => server.listen(0, "127.0.0.1", resolve)),
+          ),
+          () => Effect.promise(() => new Promise<void>((resolve) => server.close(() => resolve()))),
+        );
+        const address = server.address();
+        if (!address || typeof address === "string") return yield* Effect.die("Missing HTTP port");
+        yield* Effect.gen(function* () {
+          const index = yield* Index.make("workspace");
+          const initial = yield* index.list();
+          const other = yield* Index.make("another-workspace");
+          expect((yield* other.list()).entries).toEqual(initial.entries);
+          const blocked = yield* index.search("block", 1).pipe(Effect.exit, Effect.forkChild);
+          yield* Deferred.await(control.blocked);
 
-it.effect("filters image searches before applying the result limit", () =>
+          const health = yield* HttpClient.get(`http://127.0.0.1:${address.port}`).pipe(
+            Effect.flatMap((response) => response.text),
+            Effect.provide(NodeHttpClient.layerNodeHttp),
+          );
+          expect(health).toBe("healthy");
+          yield* TestClock.adjust("1 second");
+          const queued = yield* other.list().pipe(Effect.forkChild);
+
+          yield* TestClock.adjust("19 seconds");
+          const result = yield* Fiber.join(blocked);
+          expect(Exit.isFailure(result)).toBe(true);
+          yield* Deferred.await(control.exited);
+          const recovered = yield* Fiber.join(queued);
+          expect(recovered.entries).not.toEqual(initial.entries);
+          expect((yield* index.list()).entries).toEqual(recovered.entries);
+        }).pipe(Effect.provideService(HostProcessEnvironment, control.environment));
+      }).pipe(withWorker),
+    ),
+);
+
+it.effect("cancels a blocked request before allowing the queued request to rebuild", () =>
   Effect.scoped(
     Effect.gen(function* () {
-      const items = [
-        ...Array.from({ length: 200 }, (_, index) => fileItem(`src/file-${index}.ts`)),
-        fileItem("public/icon.svg"),
-      ];
-      const fileSearch = vi.fn(() => ({
-        ok: true as const,
-        value: {
-          items,
-          scores: [],
-          totalMatched: items.length,
-          totalFiles: items.length,
-        },
-      }));
-      const finder = {
-        destroy: vi.fn(),
-        waitForIndexReady: vi.fn(async () => ({ ok: true as const, value: true })),
-        fileSearch,
-      } as unknown as FileFinder;
-      vi.spyOn(FileFinder, "create").mockReturnValueOnce({ ok: true, value: finder });
-
-      const searchIndex = yield* WorkspaceSearchIndex.make("/workspace/project");
-      const resultWithoutKind = yield* searchIndex.search("", 200, undefined, true);
-      const resultWithDirectoryKind = yield* searchIndex.search("", 200, "directory", true);
-
-      expect(resultWithoutKind.entries).toEqual([{ kind: "file", path: "public/icon.svg" }]);
-      expect(resultWithDirectoryKind.entries).toEqual([{ kind: "file", path: "public/icon.svg" }]);
-      expect(fileSearch).toHaveBeenCalledTimes(2);
-      expect(fileSearch).toHaveBeenCalledWith("", { pageSize: 25_002 });
-    }),
+      const control = yield* receipts();
+      yield* Effect.gen(function* () {
+        const index = yield* Index.make("workspace");
+        const original = yield* index.list();
+        const blocked = yield* index.search("block", 1).pipe(Effect.forkChild);
+        yield* Deferred.await(control.blocked);
+        const queued = yield* index.list().pipe(Effect.forkChild);
+        yield* Fiber.interrupt(blocked);
+        yield* Deferred.await(control.exited);
+        expect((yield* Fiber.join(queued)).entries).not.toEqual(original.entries);
+      }).pipe(Effect.provideService(HostProcessEnvironment, control.environment));
+    }).pipe(withWorker),
   ),
 );
 
-it.effect("preserves unexpected FileFinder creation failures", () =>
-  Effect.gen(function* () {
-    const cause = new Error("native initialization failed");
-    vi.spyOn(FileFinder, "create").mockImplementationOnce(() => {
-      throw cause;
-    });
-
-    const error = yield* Effect.flip(
-      Effect.scoped(WorkspaceSearchIndex.make("/workspace/project")),
-    );
-
-    expect(error).toMatchObject({
-      _tag: "WorkspaceSearchIndexCreateFailed",
-      cwd: "/workspace/project",
-      reason: "FileFinder.create threw unexpectedly.",
-      cause,
-    });
-  }),
-);
-
-it.effect("keeps returned FileFinder creation diagnostics out of the cause chain", () =>
-  Effect.gen(function* () {
-    vi.spyOn(FileFinder, "create").mockReturnValueOnce({
-      ok: false,
-      error: "native index rejected the directory",
-    });
-
-    const error = yield* Effect.flip(
-      Effect.scoped(WorkspaceSearchIndex.make("/workspace/project")),
-    );
-
-    expect(error).toMatchObject({
-      _tag: "WorkspaceSearchIndexCreateFailed",
-      cwd: "/workspace/project",
-      reason: "native index rejected the directory",
-    });
-    expect(error.cause).toBeUndefined();
-  }),
-);
-
-it.effect("waits for the full content index warmup before returning", () =>
-  Effect.gen(function* () {
-    const waitForIndexReady = vi.fn(async () => ({ ok: true as const, value: true }));
-    const finder = {
-      destroy: vi.fn(),
-      waitForIndexReady,
-    } as unknown as FileFinder;
-    vi.spyOn(FileFinder, "create").mockReturnValueOnce({ ok: true, value: finder });
-
-    yield* Effect.scoped(WorkspaceSearchIndex.make("/workspace/project", "content"));
-
-    expect(waitForIndexReady).toHaveBeenCalledWith(15_000);
-  }),
-);
-
-it.effect("preserves a full-index warmup timeout as a structured error", () =>
-  Effect.gen(function* () {
-    const finder = {
-      destroy: vi.fn(),
-      waitForIndexReady: vi.fn(async () => ({ ok: true as const, value: false })),
-    } as unknown as FileFinder;
-    vi.spyOn(FileFinder, "create").mockReturnValueOnce({ ok: true, value: finder });
-
-    const error = yield* Effect.flip(
-      Effect.scoped(WorkspaceSearchIndex.make("/workspace/project", "content")),
-    );
-
-    expect(error).toMatchObject({
-      _tag: "WorkspaceSearchIndexScanTimedOut",
-      cwd: "/workspace/project",
-      timeout: "15 seconds",
-    });
-  }),
-);
-
-it.effect("preserves FileFinder destroy failures as structured defects", () =>
-  Effect.gen(function* () {
-    const cause = new Error("native destroy failed");
-    const finder = {
-      destroy: vi.fn(() => {
-        throw cause;
-      }),
-      waitForIndexReady: vi.fn(async () => ({ ok: true as const, value: true })),
-    } as unknown as FileFinder;
-    vi.spyOn(FileFinder, "create").mockReturnValueOnce({ ok: true, value: finder });
-
-    const exit = yield* Effect.scoped(WorkspaceSearchIndex.make("/workspace/project")).pipe(
-      Effect.exit,
-    );
-
-    expect(Exit.isFailure(exit)).toBe(true);
-    if (Exit.isFailure(exit)) {
-      expect(Cause.hasDies(exit.cause)).toBe(true);
-      const error = Cause.squash(exit.cause);
-      expect(error).toBeInstanceOf(WorkspaceSearchIndex.WorkspaceSearchIndexDestroyFailed);
-      expect(error).toMatchObject({
-        _tag: "WorkspaceSearchIndexDestroyFailed",
-        cwd: "/workspace/project",
-        cause,
-      });
-    }
-  }),
-);
-
-it.effect("preserves search and refresh failures with operation context", () =>
+it.effect("bounds initialization even when the native scan never returns", () =>
   Effect.scoped(
     Effect.gen(function* () {
-      const searchCause = new Error("native search failed");
-      const refreshCause = new Error("native scan failed");
-      const contentSearchCause = new Error("native grep failed");
-      const finder = {
-        destroy: vi.fn(),
-        waitForIndexReady: vi.fn(async () => ({ ok: true as const, value: true })),
-        mixedSearch: vi.fn(() => {
-          throw searchCause;
-        }),
-        grep: vi.fn(() => {
-          throw contentSearchCause;
-        }),
-        scanFiles: vi.fn(() => {
-          throw refreshCause;
-        }),
-      } as unknown as FileFinder;
-      vi.spyOn(FileFinder, "create").mockReturnValueOnce({ ok: true, value: finder });
-
-      const searchIndex = yield* WorkspaceSearchIndex.make("/workspace/project");
-      const query = "authorization: Bearer secret-token";
-      const searchError = yield* Effect.flip(searchIndex.search(query, 3));
-      const contentSearchError = yield* Effect.flip(
-        searchIndex.searchContents({
-          query,
-          limit: 3,
-          caseSensitive: false,
-          wholeWord: false,
-          useRegex: false,
-        }),
+      const control = yield* receipts();
+      const startup = yield* Index.make("block-initialize").pipe(
+        Effect.provideService(HostProcessEnvironment, control.environment),
+        Effect.exit,
+        Effect.forkChild,
       );
-      const refreshError = yield* Effect.flip(searchIndex.refresh());
-
-      expect(searchError).toMatchObject({
-        _tag: "WorkspaceSearchIndexSearchFailed",
-        cwd: "/workspace/project",
-        queryLength: query.length,
-        pageSize: 4,
-        reason: "FileFinder.mixedSearch threw unexpectedly.",
-        cause: searchCause,
-      });
-      expect(searchError).not.toHaveProperty("query");
-      expect(searchError.message).not.toMatch(/Bearer|secret-token/);
-      expect(contentSearchError).toMatchObject({
-        _tag: "WorkspaceSearchIndexSearchFailed",
-        cwd: "/workspace/project",
-        queryLength: query.length,
-        pageSize: 3,
-        reason: "FileFinder.grep threw unexpectedly.",
-        cause: contentSearchCause,
-      });
-      expect(contentSearchError).not.toHaveProperty("query");
-      expect(contentSearchError.message).not.toMatch(/Bearer|secret-token/);
-      expect(refreshError).toMatchObject({
-        _tag: "WorkspaceSearchIndexRefreshFailed",
-        cwd: "/workspace/project",
-        reason: "FileFinder.scanFiles threw unexpectedly.",
-        cause: refreshCause,
-      });
-    }),
+      yield* Deferred.await(control.blocked);
+      yield* TestClock.adjust("20 seconds");
+      expect(Exit.isFailure(yield* Fiber.join(startup))).toBe(true);
+      yield* Deferred.await(control.exited);
+    }).pipe(withWorker),
   ),
 );
 
-it.effect("keeps returned search diagnostics out of the cause chain", () =>
+it.effect("kills a blocked native call when the owning index scope closes", () =>
   Effect.scoped(
     Effect.gen(function* () {
-      const finder = {
-        destroy: vi.fn(),
-        waitForIndexReady: vi.fn(async () => ({ ok: true as const, value: true })),
-        mixedSearch: vi.fn(() => ({ ok: false, error: "native query rejected" })),
-        scanFiles: vi.fn(() => ({ ok: false, error: "native refresh rejected" })),
-      } as unknown as FileFinder;
-      vi.spyOn(FileFinder, "create").mockReturnValueOnce({ ok: true, value: finder });
-
-      const searchIndex = yield* WorkspaceSearchIndex.make("/workspace/project");
-      const query = "authorization: Bearer secret-token";
-      const searchError = yield* Effect.flip(searchIndex.search(query, 3));
-      const refreshError = yield* Effect.flip(searchIndex.refresh());
-
-      expect(searchError).toMatchObject({
-        _tag: "WorkspaceSearchIndexSearchFailed",
-        cwd: "/workspace/project",
-        queryLength: query.length,
-        pageSize: 4,
-        reason: "native query rejected",
-      });
-      expect(searchError).not.toHaveProperty("query");
-      expect(searchError.message).not.toMatch(/Bearer|secret-token/);
-      expect(searchError.cause).toBeUndefined();
-      expect(refreshError).toMatchObject({
-        _tag: "WorkspaceSearchIndexRefreshFailed",
-        cwd: "/workspace/project",
-        reason: "native refresh rejected",
-      });
-      expect(refreshError.cause).toBeUndefined();
-    }),
-  ),
-);
-
-it.effect("continues whole-word searches after a filtered grep page", () =>
-  Effect.scoped(
-    Effect.gen(function* () {
-      const nextCursor = {
-        __brand: "GrepCursor",
-        _offset: 1,
-      } as GrepCursor;
-      const grepResult = (
-        lineContent: string,
-        matchRanges: Array<[number, number]>,
-        cursor: GrepCursor | null,
-      ): GrepResult => ({
-        items: [
-          {
-            relativePath: "src/words.ts",
-            fileName: "words.ts",
-            gitStatus: "unmodified",
-            size: lineContent.length,
-            modified: 0,
-            isBinary: false,
-            totalFrecencyScore: 0,
-            accessFrecencyScore: 0,
-            modificationFrecencyScore: 0,
-            lineNumber: 1,
-            col: 0,
-            byteOffset: 0,
-            lineContent,
-            matchRanges,
-          },
-        ],
-        totalMatched: 1,
-        totalFilesSearched: 1,
-        totalFiles: 1,
-        filteredFileCount: 1,
-        nextCursor: cursor,
-      });
-      const grep = vi.fn((_query: string, options?: GrepOptions) =>
-        options?.cursor
-          ? { ok: true as const, value: grepResult("needle", [[0, 6]], null) }
-          : {
-              ok: true as const,
-              value: grepResult("needleSuffix", [[0, 6]], nextCursor),
-            },
+      const control = yield* receipts();
+      const scope = yield* Scope.make();
+      yield* Effect.addFinalizer(() => Scope.close(scope, Exit.void));
+      const index = yield* Index.make("workspace").pipe(
+        Scope.provide(scope),
+        Effect.provideService(HostProcessEnvironment, control.environment),
       );
-      const finder = {
-        destroy: vi.fn(),
-        waitForIndexReady: vi.fn(async () => ({ ok: true as const, value: true })),
-        grep,
-      } as unknown as FileFinder;
-      vi.spyOn(FileFinder, "create").mockReturnValueOnce({ ok: true, value: finder });
+      const blocked = yield* index.search("block", 1).pipe(Effect.exit, Effect.forkChild);
+      yield* Deferred.await(control.blocked);
+      const closing = yield* Scope.close(scope, Exit.void).pipe(Effect.forkChild);
+      yield* TestClock.adjust("1 second");
+      yield* Fiber.join(closing);
+      yield* Deferred.await(control.exited);
+      expect(Exit.isFailure(yield* Fiber.join(blocked))).toBe(true);
+    }).pipe(withWorker),
+  ),
+);
 
-      const searchIndex = yield* WorkspaceSearchIndex.make("/workspace/project", "content");
-      const result = yield* searchIndex.searchContents({
-        query: "needle",
-        limit: 1,
-        caseSensitive: true,
-        wholeWord: true,
-        useRegex: false,
+it.effect("preserves native timeout and refresh errors across IPC and recovers after a crash", () =>
+  Effect.scoped(
+    Effect.gen(function* () {
+      expect(yield* Effect.flip(Index.make("scan-timeout"))).toBeInstanceOf(
+        Index.WorkspaceSearchIndexScanTimedOut,
+      );
+      const index = yield* Index.make("workspace");
+      const original = yield* index.list();
+      expect(yield* Effect.flip(index.search("crash", 1))).toBeInstanceOf(
+        Index.WorkspaceSearchIndexSearchFailed,
+      );
+      expect((yield* index.list()).entries).not.toEqual(original.entries);
+      expect(yield* Effect.flip(index.refresh())).toMatchObject({
+        _tag: "WorkspaceSearchIndexRefreshFailed",
+        reason: "scan failed",
       });
+      expect((yield* index.list()).entries).toHaveLength(1);
+    }).pipe(withWorker),
+  ),
+);
 
-      expect(result).toEqual({
-        matches: [
-          {
-            path: "src/words.ts",
-            lineNumber: 1,
-            lineContent: "needle",
-            matchRanges: [{ start: 0, end: 6 }],
-          },
-        ],
-        truncated: false,
-      });
-      expect(grep).toHaveBeenCalledTimes(2);
-      expect(grep.mock.calls[1]?.[1]?.cursor).toBe(nextCursor);
-    }),
+it.effect("bounds native disposal and lazily rebuilds the surviving workspace", () =>
+  Effect.scoped(
+    Effect.gen(function* () {
+      const control = yield* receipts();
+      yield* Effect.gen(function* () {
+        const scope = yield* Scope.make();
+        yield* Effect.addFinalizer(() => Scope.close(scope, Exit.void));
+        const retiring = yield* Index.make("block-dispose").pipe(Scope.provide(scope));
+        const surviving = yield* Index.make("surviving");
+        const before = yield* surviving.list();
+        expect((yield* retiring.list()).entries).toEqual(before.entries);
+        const closing = yield* Scope.close(scope, Exit.void).pipe(Effect.forkChild);
+        yield* Deferred.await(control.blocked);
+        yield* TestClock.adjust("1 second");
+        yield* Fiber.join(closing);
+        yield* Deferred.await(control.exited);
+        expect((yield* surviving.list()).entries).not.toEqual(before.entries);
+        expect(yield* Effect.flip(retiring.list())).toBeInstanceOf(
+          Index.WorkspaceSearchIndexSearchFailed,
+        );
+      }).pipe(Effect.provideService(HostProcessEnvironment, control.environment));
+    }).pipe(withWorker),
+  ),
+);
+
+it.effect("releases one index without restarting the process used by another index", () =>
+  Effect.scoped(
+    Effect.gen(function* () {
+      const scope = yield* Scope.make();
+      yield* Effect.addFinalizer(() => Scope.close(scope, Exit.void));
+      const retiring = yield* Index.make("retiring").pipe(Scope.provide(scope));
+      const surviving = yield* Index.make("surviving");
+      const before = yield* surviving.list();
+      expect((yield* retiring.list()).entries).toEqual(before.entries);
+      yield* Scope.close(scope, Exit.void);
+      expect((yield* surviving.list()).entries).toEqual(before.entries);
+      expect(yield* Effect.flip(retiring.list())).toBeInstanceOf(
+        Index.WorkspaceSearchIndexSearchFailed,
+      );
+      expect((yield* surviving.list()).entries).toEqual(before.entries);
+    }).pipe(withWorker),
+  ),
+);
+
+it.effect("cancels a queued request without terminating the active search process", () =>
+  Effect.scoped(
+    Effect.gen(function* () {
+      const control = yield* receipts();
+      yield* Effect.gen(function* () {
+        const index = yield* Index.make("workspace");
+        const other = yield* Index.make("other");
+        const blocked = yield* index.search("block", 1).pipe(Effect.forkChild);
+        yield* Deferred.await(control.blocked);
+        const queued = yield* other.list().pipe(Effect.forkChild);
+        yield* Fiber.interrupt(queued);
+        expect(yield* Deferred.isDone(control.exited)).toBe(false);
+        yield* Fiber.interrupt(blocked);
+        yield* Deferred.await(control.exited);
+        expect((yield* other.list()).entries).toHaveLength(1);
+      }).pipe(Effect.provideService(HostProcessEnvironment, control.environment));
+    }).pipe(withWorker),
   ),
 );
