+++ apps/server/src/workspace/WorkspaceSearchIndexService.ts
@@ -1,0 +1,104 @@
+import * as Context from "effect/Context";
+import type * as Effect from "effect/Effect";
+import * as Schema from "effect/Schema";
+import type {
+  ProjectEntryKind,
+  ProjectListEntriesResult,
+  ProjectSearchContentsInput,
+  ProjectSearchContentsResult,
+  ProjectSearchEntriesResult,
+} from "@t3tools/contracts";
+
+export class WorkspaceSearchIndexCreateFailed extends Schema.TaggedError<WorkspaceSearchIndexCreateFailed>()(
+  "WorkspaceSearchIndexCreateFailed",
+  {
+    cwd: Schema.String,
+    reason: Schema.String,
+    cause: Schema.optional(Schema.Defect()),
+  },
+) {
+  override get message(): string {
+    return `Failed to create the workspace search index for '${this.cwd}'.`;
+  }
+}
+
+export class WorkspaceSearchIndexScanTimedOut extends Schema.TaggedError<WorkspaceSearchIndexScanTimedOut>()(
+  "WorkspaceSearchIndexScanTimedOut",
+  {
+    cwd: Schema.String,
+    timeout: Schema.String,
+  },
+) {
+  override get message(): string {
+    return `Workspace search index for '${this.cwd}' did not finish scanning within ${this.timeout}`;
+  }
+}
+
+export class WorkspaceSearchIndexSearchFailed extends Schema.TaggedError<WorkspaceSearchIndexSearchFailed>()(
+  "WorkspaceSearchIndexSearchFailed",
+  {
+    cwd: Schema.String,
+    queryLength: Schema.Number,
+    pageSize: Schema.Number,
+    reason: Schema.String,
+    cause: Schema.optional(Schema.Defect()),
+  },
+) {
+  override get message(): string {
+    return `Workspace search failed for '${this.cwd}'.`;
+  }
+}
+
+export class WorkspaceSearchIndexRefreshFailed extends Schema.TaggedError<WorkspaceSearchIndexRefreshFailed>()(
+  "WorkspaceSearchIndexRefreshFailed",
+  {
+    cwd: Schema.String,
+    reason: Schema.String,
+    cause: Schema.optional(Schema.Defect()),
+  },
+) {
+  override get message(): string {
+    return `Failed to refresh the workspace search index for '${this.cwd}'.`;
+  }
+}
+
+export class WorkspaceSearchIndexDestroyFailed extends Schema.TaggedError<WorkspaceSearchIndexDestroyFailed>()(
+  "WorkspaceSearchIndexDestroyFailed",
+  {
+    cwd: Schema.String,
+    cause: Schema.Defect(),
+  },
+) {
+  override get message(): string {
+    return `Failed to destroy the workspace search index for '${this.cwd}'.`;
+  }
+}
+
+export type WorkspaceSearchIndexError =
+  | WorkspaceSearchIndexCreateFailed
+  | WorkspaceSearchIndexScanTimedOut
+  | WorkspaceSearchIndexSearchFailed
+  | WorkspaceSearchIndexRefreshFailed;
+
+export class WorkspaceSearchIndex extends Context.Service<
+  WorkspaceSearchIndex,
+  {
+    readonly list: () => Effect.Effect<ProjectListEntriesResult, WorkspaceSearchIndexSearchFailed>;
+    readonly search: (
+      query: string,
+      limit: number,
+      kind?: ProjectEntryKind,
+      imageOnly?: boolean,
+    ) => Effect.Effect<ProjectSearchEntriesResult, WorkspaceSearchIndexSearchFailed>;
+    readonly searchContents: (
+      input: Omit<ProjectSearchContentsInput, "cwd">,
+    ) => Effect.Effect<ProjectSearchContentsResult, WorkspaceSearchIndexSearchFailed>;
+    readonly refresh: () => Effect.Effect<
+      void,
+      WorkspaceSearchIndexRefreshFailed | WorkspaceSearchIndexScanTimedOut
+    >;
+  }
+>()("t3/workspace/WorkspaceSearchIndexService/WorkspaceSearchIndex") {}
+
+export const WORKSPACE_SEARCH_INDEX_VARIANTS = ["paths", "content"] as const;
+export type WorkspaceSearchIndexVariant = (typeof WORKSPACE_SEARCH_INDEX_VARIANTS)[number];
