+++ apps/server/vite.config.ts
@@ -36,7 +36,7 @@ export default mergeConfig(
       },
     },
     pack: {
-      entry: ["src/bin.ts", "src/claudeHistoryWorker.ts"],
+      entry: ["src/bin.ts", "src/claudeHistoryWorker.ts", "src/workspaceSearchWorker.ts"],
       outDir: "dist",
       sourcemap: true,
       clean: true,
